New UI system
Updated UI system

It now supports controller and keyboard inputs. The widget for the in game menu has to be updated for it to start working in an actual level.

Logic is there and I tried to comment on everything (in Main Menu Level, I can explain how it works if needed)